Finished up a few more cases for the tester. Realized I definitely should have started on this earlier. Final version of the Beta since I need to study for a test Monday morning. Still have more cases to finish up.

+++ tokenize.cpp
@@ -2,7 +2,7 @@
 
 token StringFormat(token Unformatted)
 {
-	for (int i = 0; i < Unformatted.size(); i++)
+	for (size_t i = 0; i < Unformatted.size(); i++)
 	{
 		if (Unformatted.at(i) == '(')
 		{
